Extract `misc/rendered_image.h` out of `render.h`
