+++ src/capture/image/image_scaler.h
@@ -1,4 +1,4 @@
-// SPDX-FileCopyrightText:  2023-2025 The DOSBox Staging Team
+// SPDX-FileCopyrightText:  2023-2026 The DOSBox Staging Team
 // SPDX-License-Identifier: GPL-2.0-or-later
 
 #ifndef DOSBOX_IMAGE_SCALER_H
@@ -8,7 +8,7 @@
 
 #include "image_decoder.h"
 
-#include "gui/render/render.h"
+#include "misc/rendered_image.h"
 #include "utils/rgb888.h"
 
 enum class OutputPixelFormat { Indexed8, Rgb888 };
